migrate known-busted symbols to own file
